fixed dockable window bug
the react-dockable plugin used the window object all over it. So to solve this problem I pulled the code directly into the project and built a safeWindow object to overide it. this should fix server builds. and in the future i can rewrite the docker code to be modern and work better.