fixed annual timers not showing end message
timer was showing no message due to a return message and fixed that by removing it and putting the rest in an else. also fixed the end message showing when days hit 0 but there was still hours left.
i added that hours, minutes and seconds had to reach zero then the message will hit. days are 0 so when the day is  no longer zero or hits -1 it will then jump to next year.