fixed style bug that i made with the last few changes yesterday
I fixed the styling bug i made from yesterday when editing the list-menu-items-grid. I removed this class from the document view as it should not have been there. I added a new property pageName for the layout so i can inject it into the class so that css page targeting can be done. also removed button import that was unused and creating a warning in the documents page.